ECS: Add functionnal velocity with int

+++ src/ECS/ECSCustomTypes.hpp
@@ -52,8 +52,8 @@ namespace Types {
     };
 
     struct Velocity {
-            float speedX;
-            float speedY;
+            int speedX;
+            int speedY;
 
             NLOHMANN_DEFINE_TYPE_INTRUSIVE(Velocity, speedX, speedY);
     };
